fix: delete communication links on_trash

+++ frappe/model/delete_doc.py
@@ -275,7 +275,7 @@ def delete_dynamic_links(doctype, name):
 	delete_references('Document Follow', doctype, name, 'ref_doctype', 'ref_docname')
 
 	# unlink communications
-	clear_communication_references('Communication', doctype, name)
+	clear_communication_references(doc='Communication', link_doctype=doctype, link_name=name)
 
 	clear_references('Activity Log', doctype, name)
 	clear_references('Activity Log', doctype, name, 'timeline_doctype', 'timeline_name')
@@ -296,10 +296,15 @@ def clear_references(doctype, reference_doctype, reference_name,
 			{1}=%s and {2}=%s'''.format(doctype, reference_doctype_field, reference_name_field), # nosec
 		(reference_doctype, reference_name))
 
-def clear_communication_references(doctype, link_doctype, link_name):
-	references = frappe.get_list("Dynamic Link", filters={"parent": doctype, "link_doctype": link_doctype, "link_name": link_name})
-	for reference in references:
-		frappe.delete_doc("Dynamic Link", reference.name)
+def clear_communication_references(doc, method=None, link_doctype=None, link_name=None):
+	if method:
+		link_doctype = doc.doctype
+		link_name = doc.name
+
+	references = frappe.get_list("Dynamic Link", filters={"parenttype": "Communication", "link_doctype": link_doctype, "link_name": link_name})
+	if references:
+		for reference in references:
+			frappe.delete_doc("Dynamic Link", reference.name)
 
 def insert_feed(doc):
 	from frappe.utils import get_fullname
